Switch to React ref API instead of string refs
